+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v1.21.6 (2024-04-15)
+
+### Bug Fixes and Other Changes
+
+ * Gates inherit `targets`
+ * Make `GPhase` more efficient
+ * rename shifting field to local detuning
+
+## v1.21.5 (2024-04-11)
+
+### Bug Fixes and Other Changes
+
+ * Fix a bug in the AHS local simulator when using local detuning with certain pattern with empty sites
+
 ## v1.21.4 (2024-04-10)
 
 ### Bug Fixes and Other Changes